LayoutTests/fast/forms/checkValidity-003-expected.txt

This test checks if checkValidity() returns correctly a true (meaning no error) result on form element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


 lorem ipsum
PASS f.checkValidity() is true
PASS successfullyParsed is true

TEST COMPLETE
